fix: 자산화면 패키지 이동 #5
