+++ app/src/main/res/layout/activity_connect_property.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.main.fixed.ConnectPropertyActivity"
+    tools:context=".ui.main.property.ConnectPropertyActivity"
     android:padding="20dp">
 
     <LinearLayout
