+++ app/src/main/res/layout/activity_my_property.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="20dp"
-    tools:context=".ui.main.fixed.MyPropertyActivity">
+    tools:context=".ui.main.property.MyPropertyActivity">
 
     <LinearLayout
         android:id="@+id/layout_back"
