+++ app/src/main/res/layout/activity_property_agree.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="20dp"
-    tools:context=".ui.main.fixed.PropertyAgreeActivity">
+    tools:context=".ui.main.property.PropertyAgreeActivity">
 
     <LinearLayout
         android:id="@+id/layout_back"
