+++ app/src/main/res/layout/fragment_bank.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.fixed.BankFragment"
+    tools:context=".ui.main.property.BankFragment"
     android:padding="20dp">
 
     <LinearLayout
